Added basic support for submodules
When uploading to an empty directory, submodules are skipped and user is instructed to upload the submodules manually.

On updating to a newer version, git-upload now checks whether any of the submodules were changed. If so, the changed files are uploaded.

At this point, git-upload does not support adding or removing submodules. Removals of submodules are ignored, and adding a submodule results in a warning displayed.